+++ framework/CHANGELOG.md
@@ -19,7 +19,7 @@ Yii Framework 2 Change Log
 - Bug #18648: Fix `yii\web\Request` to properly handle HTTP Basic Auth headers (olegbaturin)
 - Enh #18726: Added `yii\helpers\Json::$prettyPrint` (rhertogh)
 - Enh #18734: Added `yii\validators\EmailValidator::$enableLocalIDN` (brandonkelly)
-
+- Bug #18756: Fix `\yii\validators\ExistValidator::queryValueExists` not unique array validate (DrDeath72)
 
 2.0.42.1 May 06, 2021
 ---------------------